% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQuote.R
\name{rQuote}
\alias{rQuote}
\title{rQuote}
\usage{
rQuote(tag = "science", page_range = 1, cores = detectCores() - 1,
  OS = c("windows", "linux", "macOS"), theme = c("light", "dark"),
  popup = TRUE, control = NULL, ...)
}
\arguments{
\item{tag}{A character string specifying the desired tag to look quotes from}

\item{page_range}{Integer specifying the number of Goodreads' pages with the specific tag to look quotes from. Defaults to 1. Must be greater than zero and less or equal to 100.}

\item{cores}{Integer specifying the number of cores to be used by the function. Defaults to the number of cores dtected on the current host minus 1.}

\item{OS}{Character string specifying the Operating System of the current host. Defaults to windows.}

\item{theme}{Character string specifying whether the light or the dark theme should be used for the pop-up plot window. Defaults to 'light'.}

\item{popup}{Logical; if TRUE a pop-up plot window with the quote appears.}

\item{control}{a list of control values with components: \itemize{
\item{width: width of plotting device}
\item{height: height of plotting device}
}}
}
\description{
Prints a random quote extracted from the goodreads.com database
}
\author{
Grigorios Papageorgiou \email{g.papageorgiou@erasmusmc.nl}
}
