% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQuoteMe.R
\name{rQuoteMe}
\alias{rQuoteMe}
\title{rQuoteMe}
\usage{
rQuoteMe(quote, author, theme = "light", fonts = "Francois One",
  OS = c("windows", "linux", "macOS"), control = list(), ...)
}
\arguments{
\item{quote}{A character string specifying the quote to plot}

\item{author}{A character string specifying the author of the quote}

\item{theme}{A character string specifying whether the light or the dark theme should be used for the pop-up plot window. Defaults to 'light'.}

\item{fonts}{Character string specifying the desired Google fonts to use. Fonts are installed using function \code{\link[sysfonts]{font_add_google}} from package \strong{sysfonts}.}

\item{OS}{Character string specifying the Operating System of the current host. Defaults to windows.}

\item{control}{a list of control values with components: \itemize{
\item{width: width of plotting device in inches. Defaults to 48.}
\item{height: height of plotting device in inches. Defaults to 24.}
\item{text_size: the size of the fonts used to plot the quote in mm. Defaults to 7.}
}}
}
\description{
Plots a quote in a graphics device
}
\details{

}
\author{
Grigorios Papageorgiou \email{g.papageorgiou@erasmusmc.nl}
}
