% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQuote.R
\name{rQuote}
\alias{rQuote}
\title{rQuote}
\usage{
rQuote(tag = "science", page_range = 1, cores = detectCores() - 1,
  OS = c("windows", "linux", "macOS"), theme = c("light", "dark"),
  popup = TRUE, control = NULL, ...)
}
\arguments{
\item{tag}{A character string specifying the desired tag to look quotes from}

\item{page_range}{Integer between 0 and 100 specifying the number of goodreads' pages with the specific tag to look quotes from}
}
\description{
Prints a random quote extracted from the goodreads.com database
}
